Move `enum Chip` from fwidgen to hubtools lib.
We do some work here to preserve the existing behavior of the `is_lpc55`
function to keep from breaking API compatibility.
